cria script para calculo do indice a partir de metadados
