+++ output/metadado_tjce_4_2020.json
@@ -1,0 +1,15 @@
+{
+    "year": 2020,
+    "month": 4,
+    "agency": "tjce",
+    "has_login": false,
+    "has_captcha": false,
+    "good_url": false,
+    "data_access_option": 2,
+    "output_format": 2,
+    "has_employee_workplace": false,
+    "has_employee_role": false,
+    "base_remuneration": true,
+    "benefits": 2,
+    "discounts": 2
+}
