+++ src/indice.py
@@ -1,203 +1,81 @@
-import pandas as pd
+import json
 
 
-class Metadado:
+def calcula_indice(metadado):
+    indice = 0
 
-    has_employee_workplace = False
-    has_employee_role = False
-    base_remuneration = False
-    benefits = 0
-    discounts = 0
+    # Se não precisa de login, ganha um ponto
+    if metadado["has_login"] == False:
+        indice += 1
 
-    def __init__(
-        self,
-        year,
-        month,
-        agency,
-        has_login,
-        has_captcha,
-        good_url,
-        data_access_option,
-        output_format,
-    ):
-        self.year = year
-        self.month = month
-        self.agency = agency
-        self.has_login = has_login
-        self.has_captcha = has_captcha
-        self.good_url = good_url
-        self.data_access_option = data_access_option
-        self.output_format = output_format
+    # Se não precisa de captha, ganha um ponto
+    if metadado["has_captcha"] == False:
+        indice += 1
 
-    def __repr__(self):
-        return (
-            "<Metadado: year:%s, month:%s, agency: %s, has_login:%s, has_captcha:%s, good_url: %s, data_access_option:%s, output_format:%s, has_employee_workplace: %s, has_employee_role:%s, base_remuneration: %s, benefits:%s, discounts:%s>"
-            % (
-                self.year,
-                self.month,
-                self.agency,
-                self.has_login,
-                self.has_captcha,
-                self.good_url,
-                self.data_access_option,
-                self.output_format,
-                self.has_employee_workplace,
-                self.has_employee_role,
-                self.base_remuneration,
-                self.benefits,
-                self.discounts,
-            )
-        )
+    # Se a url é modicável via máquina, ganha um ponto
+    if metadado["good_url"] == True:
+        indice += 1
 
+    # Se tem api para acessar os dados, ganha um ponto
+    if metadado["data_access_option"] == 0:
+        indice += 1
+    # Se necessita de scraping, ganha meio ponto
+    elif metadado["data_access_option"] == 1:
+        indice += 0.5
 
-class DataAccessOption(Metadado):
-    api = 0
-    scraping = 1
-    user_emulation = 2
+    # Se o dado é disponibilizado em CSV ou ODF, ganha um ponto
+    if metadado["output_format"] == 1 or metadado["output_format"] == 5:
+        indice += 1
 
+    # Se tem dado de Lotação, ganha um ponto
+    if metadado["has_employee_workplace"] == True:
+        indice += 1
 
-class InfoCompletenessOptions(Metadado):
-    absence = 0
-    summary = 1
-    details = 2
+    # Se tem dado de Cargo, ganha um ponto
+    if metadado["has_employee_role"] == True:
+        indice += 1
 
+    # Se tem dado de Remuneração Básica, ganha um ponto
+    if metadado["base_remuneration"] == True:
+        indice += 1
 
-class OutputFormat(Metadado):
-    pdf = 0
-    odf = 1
-    xls = 2
-    json = 3
-    html = 4
-    csv = 5
+    # Se só tem o dado total de Benefícios, ganha meio ponto
+    if metadado["benefits"] == 1:
+        indice += 0.5
+    # Se detalha o dado de Benefícios, ganha um ponto
+    elif metadado["benefits"] == 2:
+        indice += 1
 
-
-def indice_overall(row, indice):
-    # Lotação
-    if row.workplace == True:
-        indice.has_employee_workplace = True
-    # Cargo
-    if row.role == True:
-        indice.has_employee_role = True
-    # Remuneração Básica
-    if row.wage == True:
-        indice.base_remuneration = True
-    # Detalhamento de Indenizações
-    if (
-        row.perks_food == True
-        or row.perks_vacation == True
-        or row.perks_transportation
-        or row.perks_pre_school == True
-        or row.perks_health == True
-        or row.perks_birth == True
-        or row.perks_housing == True
-        or row.perks_subsistence == True
-        or row.perks_compensatory_leave == True
-        or row.perks_pecuniary == True
-        or row.perks_vacation_pecuniary == True
-        or row.perks_furniture_transport == True
-        or row.perks_premium_license_pecuniary == True
-    ) and row.perks_total == True:
-        indice.benefits = InfoCompletenessOptions.details
-    elif row.perks_total == True:
-        indice.benefits = InfoCompletenessOptions.summary
-    # Detalhamento de Descontos
-    if (
-        row.discount_prev_contribution == True
-        or row.discounts_ceil_retention == True
-        or row.discounts_income_tax == True
-    ) and row.discounts_total == True:
-        indice.discounts = InfoCompletenessOptions.details
-    elif row.discounts_total == True:
-        indice.discounts = InfoCompletenessOptions.summary
+    # Se só tem o dado total de Descontos, ganha meio ponto
+    if metadado["discounts"] == 1:
+        indice += 0.5
+    # Se detalha o dado de Discontos, ganha um ponto
+    elif metadado["discounts"] == 2:
+        indice += 1
 
     return indice
 
 
-data_tj = pd.read_csv("./tjce-2020/data.csv")
-grouped_data_tj = data_tj.groupby(by=["month"]).first()
-data_mp = pd.read_csv("./mpce-2020/data.csv")
-grouped_data_mp = data_mp.groupby(by=["month"]).first()
-
-
-# Utilizei 3 meses como exemplo
-for row in grouped_data_tj.notna().itertuples():
-    if row.Index == 3:
-        metaindice_mar = Metadado(
-            2020,
-            3,
-            "tjce",
-            False,
-            False,
-            False,
-            DataAccessOption.user_emulation,
-            OutputFormat.xls,
+metadado_path = [
+    "./output/metadado_mpce_3_2020.json",
+    "./output/metadado_mpce_4_2020.json",
+    "./output/metadado_mpce_5_2020.json",
+    "./output/metadado_tjce_3_2020.json",
+    "./output/metadado_tjce_4_2020.json",
+    "./output/metadado_tjce_5_2020.json",
+]
+for i in metadado_path:
+    with open(i) as json_file:
+        data = json.load(json_file)
+        indice = calcula_indice(data)
+        pp = (
+            "Indice de transparência do "
+            + data["agency"]
+            + " em 0"
+            + str(data["month"])
+            + "/"
+            + str(data["year"])
+            + ": "
+            + str(indice)
         )
-        indice_mar = indice_overall(row, metaindice_mar)
-        print("\n TJCE - MARÇO: \n", indice_mar)
-    elif row.Index == 4:
-        metaindice_abr = Metadado(
-            2020,
-            4,
-            "tjce",
-            False,
-            False,
-            False,
-            DataAccessOption.user_emulation,
-            OutputFormat.xls,
-        )
-        indice_abr = indice_overall(row, metaindice_abr)
-        print("\n TJCE - ABRIL: \n", indice_abr)
-    elif row.Index == 5:
-        metaindice_mai = Metadado(
-            2020,
-            5,
-            "tjce",
-            False,
-            False,
-            False,
-            DataAccessOption.user_emulation,
-            OutputFormat.xls,
-        )
-        indice_mai = indice_overall(row, metaindice_mai)
-        print("\n TJCE - MAIO: \n", indice_mai)
-
-for row in grouped_data_mp.notna().itertuples():
-    if row.Index == 3:
-        metaindice_mar = Metadado(
-            2020,
-            3,
-            "mpce",
-            False,
-            False,
-            False,
-            DataAccessOption.scraping,
-            OutputFormat.xls,
-        )
-        indice_mar = indice_overall(row, metaindice_mar)
-        print("\n MPCE - MARÇO: \n", indice_mar)
-    elif row.Index == 4:
-        metaindice_abr = Metadado(
-            2020,
-            4,
-            "mpce",
-            False,
-            False,
-            False,
-            DataAccessOption.scraping,
-            OutputFormat.xls,
-        )
-        indice_abr = indice_overall(row, metaindice_abr)
-        print("\n MPCE - ABRIL: \n", indice_abr)
-    elif row.Index == 5:
-        metaindice_mai = Metadado(
-            2020,
-            5,
-            "mpce",
-            False,
-            False,
-            False,
-            DataAccessOption.scraping,
-            OutputFormat.xls,
-        )
-        indice_mai = indice_overall(row, metaindice_mai)
-        print("\n MPCE - MAIO: \n", indice_mai)
+        print(pp)
